+++ src/module/vue/components/sf-tabs/sf-notes.vue
@@ -1,26 +1,25 @@
 <template>
   <div class="flexcol" style="margen: 1rem">
-    <textarea
-      v-model="actor.data.notes"
-      @blur="save"
-    ></textarea>
+    <quill-editor v-model="actor.data.notes" />
   </div>
 </template>
 
-<style lang="less" scoped>
-input,
-textarea {
-  border-color: rgba(0, 0, 0, 0.1);
-  border-radius: 1px;
-  font-family: var(--font-primary);
-  resize: none;
-}
-</style>
-
 <script>
+import { debounce } from 'lodash'
+
 export default {
   props: {
     actor: Object,
+  },
+
+  watch: {
+    'actor.data.notes'() {
+      this.debouncedSave()
+    },
+  },
+
+  created() {
+    this.debouncedSave = debounce(this.save, 500)
   },
 
   methods: {